Basic File upload Functionality
Added a very basic file upload functionality so that there is proof of the S3 connection working. With this we are ready to merge with main so that work on all the endpoints can be done in parallel. Note that the relevant .env entries are already defined on the backend server.