Modularized Backend
Modularized the entire backend. Everything _should_ be working exactly how it was before in terms of calls/returns to the API, but now instead of having a single 1000+ line file, we have a file for each endpoint named with the REST method (i.e. GET, POST, etc.) along with the endpoint, with /-separated words camelcased and ID always referenced in all caps. For example, `/package/:id/rate` resides in "packageIDRate.ts" and is an arrow function of the same name.